move PumpClient into its own class file, move everything into a sub dir
